Set up baseurl support for GitHub Pages
This modification allows Lanyon to adapt if hosted on a GitHub Page without a
custom domain. Please refer to Jekyll's guide
http://jekyllrb.com/docs/github-pages/
for details on how it works. Basically, if your site is on a subdomain, set
baseurl in _config.yml to the subdomain. Do not include ending slashes, but do
include leading slashes. If not on a subdomain (eg you set a custom domain)
then leave baseurl as it is, ie the empty string.