Change @media queries from rem to px
Fixes an issue in some browsers where @media queries involving rem are not
calculated properly, resulting in UI quirks (eg pagination buttons stacked
vertically instead of side-by-side).

This is probably not a good change overall, but in one of my environments I
use Midori, which exhibits this problem, so I had to fix it. To see if your
browser has it as well, check http://fvsch.com/code/bugs/rem-mediaquery/