Add a simple tag list
This provides a single page which lists every tag in sorted order. Each tag's
name is an anchor and all the posts having that tag are listed beneath it in
reverse chronological order. Those posts link back to the anchors for each tag
they provide. The more common solution, having a page for each tag, is not
possible without plugins, which I am intentionally avoiding.

I also added a few sample pages so you can see tags at work. If a page has
tags, it lists them at the bottom, in sorted order.