Implement sorted sidebar entries
Entries in the sidebar are now sorted by a `pageprio` in the YAML front
matter. The pageprio has to be one digit (ie 0-9) and pages will be shown on
the sidebar in ascending order of priority (items with the same priority get
shown in the order that Jekyll parses them). Pages without a priority, or
whose priority is not in the 0-9 range, will not be shown on the sidebar.

Priorities have been added for the existing pages to keep them on the sidebar
(post list, then tag list, then about).